Configured Product to use Apache Kafka

+++ Products/src/main/java/com/example/Products/service/impl/ProductServiceImpl.java
@@ -3,17 +3,21 @@ package com.example.Products.service.impl;
 import com.example.Products.Entity.Product;
 import com.example.Products.dto.ProductDto;
 import com.example.Products.exceptions.*;
+import com.example.Products.kafka.ProductProducer;
 import com.example.Products.mapper.ProductMapper;
 import com.example.Products.repository.ProductRepository;
 import com.example.Products.security.SecurityUtils;
 import com.example.Products.service.IProductService;
+import com.example.basedomains.ProductCreatedEvent;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataAccessException;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -25,6 +29,7 @@ public class ProductServiceImpl implements IProductService {
 
     private final ProductRepository productRepository;
     private final SecurityUtils securityUtils;
+    private final ProductProducer productProducer;
 
     @Override
     public List<ProductDto> fetchProducts() {
@@ -45,7 +50,6 @@ public class ProductServiceImpl implements IProductService {
 
     @Override
     public ProductDto fetchProductById(String productId) {
-        validateProductId(productId);
         try {
             Product product = productRepository.findById(productId)
                     .orElseThrow(() -> new ResourceNotFoundException("Product", "productId", productId));
@@ -58,7 +62,6 @@ public class ProductServiceImpl implements IProductService {
 
     @Override
     public List<ProductDto> fetchProductByUserId(String userId) {
-        validateUserId(userId);
         try {
             List<Product> products = productRepository.findProductByUserId(userId).orElseThrow(
                     () -> new ResourceNotFoundException("Product", "userId", userId)
@@ -73,35 +76,44 @@ public class ProductServiceImpl implements IProductService {
     }
 
     @Override
-    public ProductDto createProduct(ProductDto productDto) {
-        //TODO: Add ownership verification
-        // 1. Validate user role
-        log.info("Creating product for user: {}", securityUtils.getCurrentUserId());
+    public ProductDto createProduct(ProductDto productDto, MultipartFile file) {
+        log.info("Received product DTO: {}", productDto);
         if (!securityUtils.getCurrentUserRole().equals("SELLER")) {
             log.warn("Unauthorized product creation attempt by user: {}", securityUtils.getCurrentUserId());
             throw new UnauthorizedAccessException("Only sellers can create products");
         }
 
-        validateProductCreationInput(productDto);
-
         try {
             Product product = ProductMapper.mapToProduct(productDto, new Product());
-
             product.setUserId(securityUtils.getCurrentUserId());
-
             Product savedProduct = productRepository.save(product);
+
+            // If a file is provided, process it
+            if (file != null && !file.isEmpty()) {
+                // Convert file bytes to a Base64-encoded string
+                String base64ImageData = encodeImageToBase64(file);
+                ProductCreatedEvent event = new ProductCreatedEvent(savedProduct.getId(), base64ImageData);
+                productProducer.sendProductCreatedEvent(event);
+            }
             return ProductMapper.mapToProductDto(savedProduct, new ProductDto());
-        } catch (DataAccessException e) {
+        } catch (IOException e) {
             log.error("Database error while creating product: {}", e.getMessage());
             throw new DatabaseOperationException("Failed to create product due to database error");
         }
     }
 
+    private String encodeImageToBase64(MultipartFile file) throws IOException {
+        try {
+            byte[] fileBytes = file.getBytes();
+            return Base64.getEncoder().encodeToString(fileBytes);
+        } catch (IOException e) {
+            log.error("Error encoding image file to Base64: {}", e.getMessage());
+            throw new InvalidInputException("Failed to process the image file");
+        }
+    }
+
     @Override
     public ProductDto updateProduct(ProductDto productDto, String productId, Authentication authentication) {
-        validateProductId(productId);
-        validateProductInput(productDto);
-
         try {
             Product existingProduct = productRepository.findById(productId).orElseThrow(
                     () -> new ResourceNotFoundException("Product", "productId", productId)
@@ -124,13 +136,13 @@ public class ProductServiceImpl implements IProductService {
 
     @Override
     public boolean deleteProduct(String productId) {
-        validateProductId(productId);
         try {
             Product product = productRepository.findById(productId).orElseThrow(
                     () -> new ResourceNotFoundException("Product", "productId", productId)
             );
-            //TODO: Add ownership verification
-
+            if(!product.getUserId().equals(securityUtils.getCurrentUserId())) {
+                throw new UnauthorizedAccessException("You are not authorized to delete this product");
+            }
 
             productRepository.deleteById(productId);
             log.info("Deleted product with id: {}", productId);
@@ -146,47 +158,5 @@ public class ProductServiceImpl implements IProductService {
         Optional.ofNullable(productDto.getDescription()).ifPresent(product::setDescription);
         Optional.ofNullable(productDto.getPrice()).ifPresent(product::setPrice);
         Optional.of(productDto.getQuantity()).ifPresent(product::setQuantity);
-    }
-
-    private void validateProductId(String productId) {
-        if (!StringUtils.hasText(productId)) {
-            throw new InvalidInputException("Product ID cannot be empty");
-        }
-    }
-
-    private void validateUserId(String userId) {
-        if (!StringUtils.hasText(userId)) {
-            throw new InvalidInputException("User ID cannot be empty");
-        }
-    }
-
-    private void validateProductInput(ProductDto productDto) {
-        if (productDto == null) {
-            throw new InvalidInputException("Product data cannot be null");
-        }
-        if (!StringUtils.hasText(productDto.getName())) {
-            throw new InvalidInputException("Product name cannot be empty");
-        }
-        if (productDto.getPrice() == null || productDto.getPrice() <= 0) {
-            throw new InvalidInputException("Invalid product price");
-        }
-    }
-
-    private void validateProductCreationInput(ProductDto productDto) {
-        if (productDto == null) {
-            throw new InvalidInputException("Product data cannot be null");
-        }
-
-        if (!StringUtils.hasText(productDto.getName())) {
-            throw new InvalidInputException("Product name cannot be empty");
-        }
-
-        if (productDto.getPrice() == null || productDto.getPrice() <= 0) {
-            throw new InvalidInputException("Product price must be a positive number");
-        }
-
-        if (productDto.getQuantity() < 0) {
-            throw new InvalidInputException("Product quantity cannot be negative");
-        }
     }
 }
